requirement.txt
django
djangorestframework
requests
django-filter
Pillow
djangorestframework_simplejwt